eudplib.eudlib.memiof -> eudplib.memio
