+++ eudplib/memio/byterw.py
@@ -5,11 +5,11 @@
 # and is released under "MIT License Agreement". Please see the LICENSE
 # file that should have been included as part of this package.
 
-from ... import core as c
-from ... import ctrlstru as cs
-from ...core.variable.evcommon import _ev
-from ...localize import _
-from ...utils import EPD, EPError, _rand_lst
+from .. import core as c
+from .. import ctrlstru as cs
+from ..core.variable.evcommon import _ev
+from ..localize import _
+from ..utils import EPD, EPError, _rand_lst
 from . import modcurpl as cp
