+++ eudplib/memio/cpbyterw.py
@@ -5,10 +5,9 @@
 # and is released under "MIT License Agreement". Please see the LICENSE
 # file that should have been included as part of this package.
 
-from eudplib import core as c
-from eudplib import ctrlstru as cs
-from eudplib import utils as ut
-
+from .. import core as c
+from .. import ctrlstru as cs
+from .. import utils as ut
 from . import modcurpl as cp
