+++ eudplib/memio/cpmemio.py
@@ -5,11 +5,11 @@
 # and is released under "MIT License Agreement". Please see the LICENSE
 # file that should have been included as part of this package.
 
-from ... import core as c
-from ... import ctrlstru as cs
-from ... import utils as ut
-from ...core.eudfunc.eudf import _EUDPredefineReturn
-from ...core.variable.evcommon import _ev
+from .. import core as c
+from .. import ctrlstru as cs
+from .. import utils as ut
+from ..core.eudfunc.eudf import _EUDPredefineReturn
+from ..core.variable.evcommon import _ev
 from . import dwepdio as dwm
 from . import modcurpl as cp
