+++ eudplib/memio/modcurpl.py
