+++ eudplib/memio/varrayreader.py
